+++ ZAPD/ZResource.h
@@ -25,6 +25,7 @@ class ZFile;
 enum class ZResourceType
 {
 	Error,
+	ActorList,
 	Animation,
 	Array,
 	AltHeader,
