+++ ZAPD/ZRoom/Commands/SetActorList.cpp
@@ -1,10 +1,10 @@
 #include "SetActorList.h"
 
+#include <cassert>
+
 #include "Globals.h"
 #include "Utils/BitConverter.h"
 #include "Utils/StringHelper.h"
-#include "ZFile.h"
-#include "ZRoom/ZNames.h"
 #include "ZRoom/ZRoom.h"
 
 SetActorList::SetActorList(ZFile* nParent) : ZRoomCommand(nParent)
@@ -16,50 +16,26 @@ void SetActorList::ParseRawData()
 	ZRoomCommand::ParseRawData();
 	numActors = cmdArg1;
 
-	uint32_t currentPtr = segmentOffset;
-
-	for (size_t i = 0; i < numActors; i++)
-	{
-		ActorSpawnEntry entry(parent->GetRawData(), currentPtr);
-
-		currentPtr += entry.GetRawDataSize();
-		actors.push_back(entry);
-	}
+	actorList = new ZActorList(parent);
+	actorList->ExtractFromBinary(segmentOffset, numActors);
 }
 
 void SetActorList::DeclareReferences(const std::string& prefix)
 {
-	if (actors.empty())
+	if (parent->HasDeclaration(segmentOffset)) {
+		delete actorList;
+		actorList = static_cast<ZActorList*>(parent->FindResource(segmentOffset));
+		assert(actorList != nullptr);
+		assert(actorList->GetResourceType() == ZResourceType::ActorList);
 		return;
-
-	std::string declaration;
-
-	size_t largestlength = 0;
-	for (const auto& entry : actors)
-	{
-		size_t actorNameLength = ZNames::GetActorName(entry.GetActorId()).size();
-		if (actorNameLength > largestlength)
-			largestlength = actorNameLength;
 	}
 
-	size_t index = 0;
-	for (auto& entry : actors)
+	if (actorList->GetName() == "")
 	{
-		entry.SetLargestActorName(largestlength);
-		declaration += StringHelper::Sprintf("\t{ %s },", entry.GetBodySourceCode().c_str());
-
-		if (index < actors.size() - 1)
-			declaration += "\n";
-
-		index++;
+		actorList->SetName(actorList->GetDefaultName(prefix));
 	}
-
-	const auto& entry = actors.front();
-
-	std::string varName = StringHelper::Sprintf("%sActorList_%06X", prefix.c_str(), segmentOffset);
-	parent->AddDeclarationArray(segmentOffset, DeclarationAlignment::Align4,
-	                            actors.size() * entry.GetRawDataSize(), entry.GetSourceTypeName(),
-	                            varName, GetActorListArraySize(), declaration);
+	actorList->DeclareVar(prefix, "");
+	parent->AddResource(actorList);
 }
 
 std::string SetActorList::GetBodySourceCode() const
@@ -70,29 +46,6 @@ std::string SetActorList::GetBodySourceCode() const
 	return StringHelper::Sprintf("SCENE_CMD_ACTOR_LIST(%i, %s)", numActors, listName.c_str());
 }
 
-size_t SetActorList::GetActorListArraySize() const
-{
-	size_t actorCount = 0;
-
-	// Doing an else-if here so we only do the loop when the game is SW97.
-	// Actor 0x22 is removed from SW97, so we need to ensure that we don't increment the actor count
-	// for it.
-	if (Globals::Instance->game == ZGame::OOT_SW97)
-	{
-		actorCount = 0;
-
-		for (const auto& entry : actors)
-			if (entry.GetActorId() != 0x22)
-				actorCount++;
-	}
-	else
-	{
-		actorCount = actors.size();
-	}
-
-	return actorCount;
-}
-
 std::string SetActorList::GetCommandCName() const
 {
 	return "SCmdActorList";
@@ -101,59 +54,4 @@ std::string SetActorList::GetCommandCName() const
 RoomCommand SetActorList::GetRoomCommand() const
 {
 	return RoomCommand::SetActorList;
-}
-
-ActorSpawnEntry::ActorSpawnEntry(const std::vector<uint8_t>& rawData, uint32_t rawDataIndex)
-{
-	actorNum = BitConverter::ToInt16BE(rawData, rawDataIndex + 0);
-	posX = BitConverter::ToInt16BE(rawData, rawDataIndex + 2);
-	posY = BitConverter::ToInt16BE(rawData, rawDataIndex + 4);
-	posZ = BitConverter::ToInt16BE(rawData, rawDataIndex + 6);
-	rotX = BitConverter::ToUInt16BE(rawData, rawDataIndex + 8);
-	rotY = BitConverter::ToUInt16BE(rawData, rawDataIndex + 10);
-	rotZ = BitConverter::ToUInt16BE(rawData, rawDataIndex + 12);
-	initVar = BitConverter::ToInt16BE(rawData, rawDataIndex + 14);
-}
-
-std::string ActorSpawnEntry::GetBodySourceCode() const
-{
-	std::string body;
-
-	std::string actorNameFmt = StringHelper::Sprintf("%%-%zus ", largestActorName + 1);
-	body =
-		StringHelper::Sprintf(actorNameFmt.c_str(), (ZNames::GetActorName(actorNum) + ",").c_str());
-
-	body += StringHelper::Sprintf("{ %6i, %6i, %6i }, ", posX, posY, posZ);
-	if (Globals::Instance->game == ZGame::MM_RETAIL)
-		body += StringHelper::Sprintf("{ SPAWN_ROT_FLAGS(%#5hX, 0x%04X)"
-		                              ", SPAWN_ROT_FLAGS(%#5hX, 0x%04X)"
-		                              ", SPAWN_ROT_FLAGS(%#5hX, 0x%04X) }, ",
-		                              (rotX >> 7) & 0b111111111, rotX & 0b1111111,
-		                              (rotY >> 7) & 0b111111111, rotY & 0b1111111,
-		                              (rotZ >> 7) & 0b111111111, rotZ & 0b1111111);
-	else
-		body += StringHelper::Sprintf("{ %#6hX, %#6hX, %#6hX }, ", rotX, rotY, rotZ);
-	body += StringHelper::Sprintf("0x%04X", initVar);
-
-	return body;
-}
-
-std::string ActorSpawnEntry::GetSourceTypeName() const
-{
-	return "ActorEntry";
-}
-
-int32_t ActorSpawnEntry::GetRawDataSize() const
-{
-	return 16;
-}
-
-uint16_t ActorSpawnEntry::GetActorId() const
-{
-	return actorNum;
-}
-
-void ActorSpawnEntry::SetLargestActorName(size_t nameSize)
-{
-	largestActorName = nameSize;
 }
